browser/components/feeds/test/test_bug589543.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=589543
-->
<head>
  <title>Test feed preview subscribe UI</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=589543">Mozilla Bug 589543</a>
<p id="display"><iframe id="testFrame" src="bug589543-data.xml"></iframe></p>
<div id="content" style="display: none">
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Bug 589543 **/
SimpleTest.waitForExplicitFinish();

addLoadEvent(function() {
  var doc = SpecialPowers.wrap($("testFrame")).contentDocument;
  var daddy = doc.getElementById("feedSubscribeLine");
  var popup = doc.getAnonymousElementByAttribute(daddy, "anonid", "handlersMenuPopup");
  isnot(popup, null, "Feed preview should have a handlers popup");
});
addLoadEvent(SimpleTest.finish);

</script>
</pre>
</body>
</html>
